# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BotStatusState = typeof BotStatusState[keyof typeof BotStatusState];


export const BotStatusState = {
  offline: 'offline',
  connecting: 'connecting',
  online: 'online',
  error: 'error',
} as const;

/**
 * @nullable
 */
export type BotStatusPosition = {
  x: number;
  y: number;
  z: number;
} | null;

export interface BotStatus {
  state: BotStatusState;
  /** @nullable */
  username: string | null;
  /** @nullable */
  host: string | null;
  /** @nullable */
  port: number | null;
  /** @nullable */
  version: string | null;
  /** @nullable */
  health: number | null;
  /** @nullable */
  position: BotStatusPosition;
  /** @nullable */
  lastEvent: string | null;
  updatedAt: string;
}

export type BotAccountAuth = typeof BotAccountAuth[keyof typeof BotAccountAuth];


export const BotAccountAuth = {
  offline: 'offline',
  microsoft: 'microsoft',
} as const;

export interface BotAccount {
  id: string;
  label: string;
  username: string;
  auth: BotAccountAuth;
  hasPassword: boolean;
  createdAt: string;
}

export type BotAccountInputAuth = typeof BotAccountInputAuth[keyof typeof BotAccountInputAuth];


export const BotAccountInputAuth = {
  offline: 'offline',
  microsoft: 'microsoft',
} as const;

export interface BotAccountInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  label: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  username: string;
  auth: BotAccountInputAuth;
  /**
     * @minLength 1
     * @maxLength 128
     * @pattern ^\S+$
     */
  password?: string;
}

export interface DeleteBotAccountResult {
  deleted: boolean;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export type BotLogLevel = typeof BotLogLevel[keyof typeof BotLogLevel];


export const BotLogLevel = {
  info: 'info',
  success: 'success',
  warning: 'warning',
  error: 'error',
  chat: 'chat',
} as const;

export interface BotLog {
  id: string;
  timestamp: string;
  level: BotLogLevel;
  message: string;
}

export interface BotPlayer {
  username: string;
  /** @nullable */
  displayName: string | null;
}

export type BotConnectInputAuth = typeof BotConnectInputAuth[keyof typeof BotConnectInputAuth];


export const BotConnectInputAuth = {
  offline: 'offline',
  microsoft: 'microsoft',
} as const;

export interface BotConnectInput {
  /** @minLength 1 */
  host: string;
  /**
     * @minimum 1
     * @maximum 65535
     */
  port: number;
  /** @minLength 1 */
  username: string;
  /** @nullable */
  version?: string | null;
  auth: BotConnectInputAuth;
  /** @nullable */
  accountId?: string | null;
}

export interface BotChatInput {
  /**
     * @minLength 1
     * @maxLength 256
     */
  message: string;
}

export interface BotCommandInput {
  /**
     * @minLength 1
     * @maxLength 256
     */
  command: string;
}

export interface BotCommandResponse {
  output: string;
  status: BotStatus;
}

export interface ErrorResponse {
  error: string;
}