# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: bot
    description: Mineflayer bot controls
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /bot/status:
    get:
      operationId: getBotStatus
      tags: [bot]
      summary: Get current bot status
      description: Returns the bot connection state, identity, server target, and recent activity.
      responses:
        "200":
          description: Bot status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatus"
  /bot/accounts:
    get:
      operationId: getBotAccounts
      tags: [bot]
      summary: List saved Minecraft accounts
      description: Returns saved account summaries without passwords.
      responses:
        "200":
          description: Saved account summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BotAccount"
    post:
      operationId: createBotAccount
      tags: [bot]
      summary: Save a Minecraft account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotAccountInput"
      responses:
        "201":
          description: Saved account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotAccount"
        "400":
          description: Invalid account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/user:
    get:
      operationId: getCurrentAuthUser
      tags: [auth]
      summary: Get the current signed-in user
      responses:
        "200":
          description: Authentication state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
  /bot/accounts/{accountId}:
    delete:
      operationId: deleteBotAccount
      tags: [bot]
      summary: Delete a saved Minecraft account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteBotAccountResult"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bot/logs:
    get:
      operationId: getBotLogs
      tags: [bot]
      summary: Get recent bot logs
      description: Returns recent connection, chat, and error events from the in-memory bot session.
      responses:
        "200":
          description: Recent logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BotLog"
  /bot/players:
    get:
      operationId: getBotPlayers
      tags: [bot]
      summary: Get players visible to the bot
      description: Returns the names currently synchronized in the Mineflayer player map.
      responses:
        "200":
          description: Visible players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BotPlayer"
  /bot/connect:
    post:
      operationId: connectBot
      tags: [bot]
      summary: Connect the Mineflayer bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotConnectInput"
      responses:
        "200":
          description: Bot connection started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatus"
        "400":
          description: Invalid bot connection settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Bot is already connecting or connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bot/disconnect:
    post:
      operationId: disconnectBot
      tags: [bot]
      summary: Disconnect the Mineflayer bot
      responses:
        "200":
          description: Bot disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatus"
  /bot/chat:
    post:
      operationId: sendBotChat
      tags: [bot]
      summary: Send a chat message through the bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotChatInput"
      responses:
        "200":
          description: Chat message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatus"
        "400":
          description: Invalid chat message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bot/command:
    post:
      operationId: runBotCommand
      tags: [bot]
      summary: Run a safe local bot command
      description: Runs one of the supported operator commands such as goto, serverlist, or help.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotCommandInput"
      responses:
        "200":
          description: Command accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotCommandResponse"
        "400":
          description: Invalid command
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Bot is not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    BotStatus:
      type: object
      properties:
        state:
          type: string
          enum: [offline, connecting, online, error]
        username:
          type: ["string", "null"]
        host:
          type: ["string", "null"]
        port:
          type: ["number", "null"]
          multipleOf: 1
        version:
          type: ["string", "null"]
        health:
          type: ["number", "null"]
        position:
          type: ["object", "null"]
          properties:
            x: { type: number }
            y: { type: number }
            z: { type: number }
          required: [x, y, z]
        lastEvent:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
      required: [state, username, host, port, version, health, position, lastEvent, updatedAt]
    BotAccount:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        username:
          type: string
        auth:
          type: string
          enum: [offline, microsoft]
        hasPassword:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, label, username, auth, hasPassword, createdAt]
    BotAccountInput:
      type: object
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 80
        username:
          type: string
          minLength: 1
          maxLength: 120
        auth:
          type: string
          enum: [offline, microsoft]
        password:
          type: string
          minLength: 1
          maxLength: 128
          pattern: '^\S+$'
      required: [label, username, auth]
    DeleteBotAccountResult:
      type: object
      properties:
        deleted:
          type: boolean
      required: [deleted]
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
      required: [id, email, firstName, lastName, profileImageUrl]
    AuthUserEnvelope:
      type: object
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
      required: [user]
    BotLog:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        level:
          type: string
          enum: [info, success, warning, error, chat]
        message:
          type: string
      required: [id, timestamp, level, message]
    BotPlayer:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: ["string", "null"]
      required: [username, displayName]
    BotConnectInput:
      type: object
      properties:
        host:
          type: string
          minLength: 1
        port:
          type: number
          multipleOf: 1
          minimum: 1
          maximum: 65535
        username:
          type: string
          minLength: 1
        version:
          type: ["string", "null"]
        auth:
          type: string
          enum: [offline, microsoft]
        accountId:
          type: ["string", "null"]
      required: [host, port, username, auth]
    BotChatInput:
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 256
      required: [message]
    BotCommandInput:
      type: object
      properties:
        command:
          type: string
          minLength: 1
          maxLength: 256
      required: [command]
    BotCommandResponse:
      type: object
      properties:
        output:
          type: string
        status:
          $ref: "#/components/schemas/BotStatus"
      required: [output, status]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]